k: Render graph
`k` is supposed to be a visualization tool, but so far we've just printed a
bunch of OIDs... Now comes the visualization part!

There's a convenient file format called "dot" that can describe a graph. This is
a textual format. We'll generate a graph of all commits and refs in dot format
and then visualize it using the "dot" utility that comes with Graphviz.

(If you're unfamiliar with dot or Graphviz please look it up online.)

The graph will contains a node for each commit, that points to the parent
commit. The graph will also contain a node for each ref, which points to the
relevant commit.

At this point, `ugit k` is fully functional and I encourage you to play with it.
Create a crazy branching history and a bunch of tags and see for yourself that
`ugit k` can draw all that visually.

%% METADATA
Id: k-render-graph