data: Create RefValue container
To make working with symbolic refs easier, we will create a `RefValue` container
to represent the value of a ref. `RefValue` will have a property `symbolic` that
will say whether it's a symbolic or a direct ref.

This change is just refactoring, we will wrap every OID that is written or read
from a ref in a `RefValue`.

%% METADATA
Id: create-refvalue