checkout: Switch branches
Finally! After this change, if we checkout a branch, we will point HEAD to the
ref of that branch (making HEAD symbolic). If we check out anything else (such
as a tag or an OID directly), we'll just point HEAD directly at the commit
OID (making HEAD non-symbolic).

We treat every ref under *refs/heads* as a branch.

The behaviour described in the introduction of branches now works. We can make
commits on a branch and it will advance along with HEAD. If we have this:

    o-----o-----o-----o-----o-----o-----O (3d8773...)
                       \                ^
                        ----o---o    branch1
                                ^
                             branch2

(Assume that the OID of the last commit is 3d8773.) If we checkout the commit by
branch name (`ugit checkout branch1`), HEAD will become a symbolic ref that
points to the ref of *branch1*. The physical value of HEAD will be `ref:
refs/heads/branch1`.

    o-----o-----o-----o-----o-----o-----O (3d8773...)
                       \                ^
                        ----o---o    branch1 <--- HEAD
                                ^
                             branch2

And if you create a commit, it will advance *branch1* together with HEAD:

    o-----o-----o-----o-----o-----o-----O-----o
                       \                      ^
                        ----o---o          branch1 <--- HEAD
                                ^
                          branch2

Note that we can checkout that same commit by its OID (`ugit checkout 3d8773`
instead of `ugit checkout branch1`). This is actually a different thing!

HEAD ultimately still points to the same commit in both cases, but if we
checkout the commit by OID , HEAD will point directly to OID *3d8773*:

    o-----o-----o-----o-----o-----o-----O (3d8773...)
                       \                ^
                        ----o---o    branch1,HEAD
                                ^
                             branch2

The physical value of HEAD will be "3d8773". Now commits will NOT advance the
*branch1*, only HEAD will advance:

    o-----o-----o-----o-----o-----o-----O-----o
                       \                ^     ^
                        ----o---o    branch1  HEAD
                                ^
                             branch2

This situation is called **detached HEAD**. You might have heard about it when
you were working with real Git. It's a confusing concept at first, but hopefully
it's easier to understand now since we've built everything ourselves.

**Detached HEAD** is a bit of a dangerous situation to be in because you can
lose some commits if you work in detached HEAD and then switch to another
branch. For example imagine you do `ugit checkout branch2`:

    o-----o-----o-----o-----o-----o-----O-----o
                       \                ^
                        ----o---o    branch1
                                ^
                             branch2 <--- HEAD

As you can see, no refs point to the latest commit that we've previousy made on
top of *branch1*. You'll need to know its OID in order to retrieve it again,
which is currently problematic if you don't remember it.

So make sure you're never in detached HEAD state unless you intend to.

%% METADATA
Id: checkout-switch-branches