init: Set HEAD to master on init
We will make HEAD point to a ref *refs/heads/master* so that the repository has
an initial branch, called *master* (meaning the main branch). The branch could
have been named anything we'd like but *master* is the standard name used for
the first branch in Git.

Without this change, a fresh ugit repository would have started in detached
HEAD, but now we'll have branch right from the start to save the user some
typing and confusion.

Note that now we'll have 2 init functions: One in *base.py* and one in *data.py*.
`base.init()` calls `data.init()` and each initializes a different aspect of the
repository. As you remember, *data.py* handles everything that directly touches
the disk (object database and refs) and *base.py* handles the higher-level
concepts built on top of *data.py*. Because "setting an initial branch" is a
higher-level concept, its initialization belongs in *base.py*.

%% METADATA
Id: init-head-to-master