show: Print commit message
So far, from the user's perspective, commit objects are opaque. The user can see
the commit message using `log` or `cat-file`, and maybe the user can read the
tree object associated with the commit, but that's basically it.

Most importantly, the user can't see the changes between commits (also known as
diffs). As you saw, Git stores the snapshot of the entire directory for each
commit, not the diffs.

Seeing diffs is a very important feature in a version control system, since
usually when reviewing commits the diff is the part you use to judge the content
of the commit.

We will add a `ugit show` CLI command which will print a commit object for us in
a useful way - the commit message and the textual diff from the last commit.

In this change, let's create a new CLI command for `show`, and just print the
commit message to get started. We will print it in the same format as `log`, so
let's refactor this to `_print_commit`.

(Note: in Git, `git show` can show various types of objects. In ugit, `ugit
show` will just work for commit objects, to simplify the code)

%% METADATA
Id: show-commit-message