show: List changed files in commit
Let's first figure out which files were changed by the commit.

We will compare the tree of the commit to the tree of its parent commit. If a
file has one OID in the former tree and a different OID in the latter tree, it
means that the file was changed (identical files would have had the same OID,
since the OID is a hash over the content).

Let's create a new module - *diff.py*. This module will contain the code that
deals with computing differences between objects. We will implement an important
function called `compare_trees()`, this function will take a list of trees and
will return them grouped by filename. This way, for each file we can get all its
OIDs in the different trees.

Using `diff.compare_trees()`, the task of displaying all changed files is
straightforward. We will create `diff.diff_trees()` that takes two trees,
compares them and return all entries that have different OIDs. Then what is left
is to call `diff_trees()` from `show` with the correct trees (of the commit we
want to show and its parent).

%% METADATA
Id: show-changed-files