push: Don't allow force push
Our current `push` implementation isn't safe from a certain kind of race. If two
people are working on the same branch that is pushed to a common remote
repository, they might overwrite each other's work. Consider this case:

- My repository:

                      refs/heads/master
                      v
      o---o---o---o---@

- My colleague's repository:

                      refs/heads/master
                      v
      o---o---o---o---$

- Remote repository:

                  refs/heads/master
                  v
      o---o---o---o

Note that both me and my colleague are working with the same remote and each of
us made a different commit on top of the same commit from the remote repository.
Now, if my colleague will `push`, the state will look like

- My repository:

                      refs/heads/master
                      v
      o---o---o---o---@

- My colleague's repository:

                      refs/heads/master
                      v
      o---o---o---o---$

- Remote repository:

                      refs/heads/master
                      v
      o---o---o---o---$

If I don't pay attention and also run `push`, ugit will happily overwrite my
colleague's latest commit, like this:

- My repository:

                      refs/heads/master
                      v
      o---o---o---o---@

- My colleague's repository:

                      refs/heads/master
                      v
      o---o---o---o---$

- Remote repository:

                      refs/heads/master
                      v
      o---o---o---o---@

To prevent it from happening we're going to allow pushing only in two cases:

- The ref that we're pushing doesn't exist yet on the remote. It means that it's
  a new branch and there is no risk of overwriting other's work.

- If the remote ref does exist, it must point to a commit that is an ancestor of
the pushed ref. This ancestry means that the local commit is based on the remote
commit, which means that the remote commit not getting overwritten, since it's
part of the history of the newly pushed commit.

Let's return the the previous example. Assuming the situtation looks like this:

- My repository:

                      refs/heads/master
                      v
      o---o---o---o---@

- My colleague's repository:

                      refs/heads/master
                      v
      o---o---o---o---$

- Remote repository:

                      refs/heads/master
                      v
      o---o---o---o---$

If I try to run `ugit push master`, the new assert will fire, because the `$`
commit isn't a ancestor of the `@` commit. To reconcile, I need to run `ugit
fetch`, which will retrieve the latest master:

- My repository:

                      refs/heads/master
                      v
      o---o---o---o---@
                   \
                    --$
                      ^
                      refs/remote/master

- Remote repository:

                      refs/heads/master
                      v
      o---o---o---o---$

Then I need to run `ugit merge remote/master`, which will make combine the
branches:

- My repository:

                          refs/heads/master
                          v
      o---o---o---o---@---o
                   \     /
                    --$--
                      ^
                      refs/remote/master

- Remote repository:

                      refs/heads/master
                      v
      o---o---o---o---$

Only now I can run `push` successfully, since the `$` commit which the remote
points to is an ancestor of master. After `push`:

- My repository:

                          refs/heads/master
                          v
      o---o---o---o---@---o < refs/remote/master
                   \     /
                    --$--

- Remote repository:

                          refs/heads/master
                          v
      o---o---o---o---@---o
                   \     /
                    --$--

BTW, real Git has a CLI option `--force` to disable this safety check, but we'll
skip it here.

%% METADATA
Id: push-dont-allow-force