add: Record added files in the index
Right now, when we run `ugit commit`, all the changes in the working directory
are written to the next commit. There are situations where this isn't very
convenient. For example, we could have made multiple unrelated changes during
the course of our work and we don't necessarily want to group them in the same
commit.

We will now implement a feature called **the index** which will allow finer
grained control over commited files. It will allow us to specify which files in
the working directory should be part of the commit and which shouldn't.

The flow will look like this:

- The user modifies some files in the working directory.

- For each file that the user wishes to commit, he will run `ugit add
path/to/file`. ugit will put the file into the object database and remember its
OID in **the index**. The index is a dictionary that maps filenames to their
last remembered OID. The index will be saved as a JSON file in the *.ugit*
directory so that we can persist it between invocations of ugit.

- The user runs `ugit commit`, at which point `write-tree` will take the
contents of **the index** and write it to a tree object. Since the index maps
file names to OIDs, just like a tree object, the conversion will be easy.

Previosuly, `write-tree` would take its data from the working directory. Now it
will take the data from the index, and there would be a separate `add` command
to put data into the index. This effectively allows the user to control what
goes into the next commit by adding ("staging" in Git's lingo) the relevant
files to the index.

In this change we're going to lay out the foundations for the index. We will add
a brand new `add` command to the CLI, then we will add a function called
`get_index` that can read and write the index in JSON format, and then once
`add` is invoked on a file it will record the current state of the file in the
index.

Note that this change doesn't affect ugit yet in any way, since `write-tree`
still takes files from the working directory. We will connect `write-tree` to
the index in a following change. For now you can run `ugit add` on a bunch of
files and make sure that they indeed appear in *.ugit/index*.

(BTW, real Git has a more complex and optimized data structure for the index,
but JSON will be enough for us to get the idea.)

%% METADATA
Id: add-files-to-index