merge: Merge in working directory
We'll continue with the previous example:

                    HEAD
                    v
    o---o---o---o---o
         \
          --o---o
                ^
                some-branch

As you remember, we want to merge HEAD and *some-branch*. Assume that both HEAD
and *some-branch* have a file named *main.py* in their trees. When merging, we
want to create a new *main.py* that will contain the content of both HEAD's
*main.py* and *some-branch*'s *main.py*.

Here's HEAD's *main.py*:

    def main ():
        print (*This functions is cool*)
        print ("It prints stuff")
        print ("It can even return a number:")
        return 7

Here's *some-branch*'s *main.py*:

    def main ():
        print ("1+1 = 2")
        print ("This functions is cool")
        print ("It prints stuff")

You can see that they have these lines in common:

    def main ():
        print ("This functions is cool")
        print ("It prints stuff")

And also each has some unique lines. When merging them together, we would
ideally arrive to a combined *main.py* that looks like this:

    def main ():
        print ("1+1 = 2")
        print ("This functions is cool")
        print ("It prints stuff")
        print ("It can even return a number:")
        return 7

As you can see the common lines are still there and we've also added the unique
lines from HEAD and *some-branch*.

The `diff` shell command that we've used earlier can actually do this merging.
The command `diff -DHEAD file1 file2` outputs merged version of file1 and file2.
If we run `diff -DHEAD` on our two versions of main.py we would get:

    def main ():
    #ifdef HEAD
        print ("1+1 = 2")
    #endif /* HEAD */
        print ("This functions is cool")
        print ("It prints stuff")
    #ifndef HEAD
        print ("It can even return a number:")
        return 7
    #endif /* ! HEAD */

Which is pretty close to what we want! (There are the "#ifdef" lines which mark
where each line came from, they add some noise but we will ignore them for now.
We will fix that later.)

The example above was for individual files. Let's write some logic to merge two
full trees together. As you remember, we had `base.read_tree()` command which
takes a tree and extract it into the working directory. We will now create
`base.read_tree_merged()` which will take two trees and extract a merged version
of them into the working directory.

We'll need a helper function that calls the `diff` shell command. For that we
will create `diff.merge_blobs()` that gets two OIDs and returns their merged
content.

Then we will create `diff.merge_trees()`, which gets two trees and in turn calls
`diff.merge_blobs()` to merge each two files in the trees, outputting one merged
tree.

`base.read_tree_merged()` just calls `diff.merge_trees()` and writes the resulting
merged tree to the working directory.

And finally, to kick things off, `base.merge()` takes the tree of the HEAD and the
tree of the branch we want to merge with and calls `base.read_tree_merged()`.

Note that at this stage we didn't create any merge commits like shown in the
diagrams, we've just took two branches and merged them in the working directory.
This isn't super useful at this stage, but it's the first step. Also, the `diff`
command adds all these #ifdefs so the merge isn't very usable at this stage, but
we will address this later.

This commit is actually pretty complicated, please pause and make sure that you
understood what's going on here.

%% METADATA
Id: merge-working-directory